Create the Boa v0.17 release post (#94)
* Create the Boa v0.17 release post

While this is still a very work-in-progress post, and we're still missing some features from reaching the `main` branch in Boa, this will be a HUGE release, so it would be great if we could already start writing the post.

If the post becomes too large, we might need to split it a bit and create multiple posts explaining some features (such as optimizations), and release them before the final release post, so that we can link back to them.

* - adds default build
- improves the spacing for wider screens
- Defers loading of mermaid.js

* Document modules example

* - Remove redirect of mermaid
- put opencollective's script as low priority (its at the bottom of the page)

* Updates on styling and content

* Added the final touches to the release post

* Update posts/2023-07-05-boa-release-17.md

Co-authored-by: José Julián Espina <jedel0124@gmail.com>

* Update posts/2023-07-05-boa-release-17.md

Co-authored-by: José Julián Espina <jedel0124@gmail.com>

* Updated dependencies & conformance

* Update posts/2023-07-08-boa-release-17.md

* Update posts/2023-07-08-boa-release-17.md

* Update posts/2023-07-08-boa-release-17.md

* Update posts/2023-07-08-boa-release-17.md

* Update posts/2023-07-08-boa-release-17.md

* Update posts/2023-07-08-boa-release-17.md

* Fixed typos and removed non-needed explanation

---------

Co-authored-by: jasonwilliams <jase.williams@gmail.com>
Co-authored-by: José Julián Espina <jedel0124@gmail.com>